refactor(lcm): split evals.py into types/seeding/answerer/retrievers/runner package
``backend/app/core/lcm/evals.py`` was 873 lines — well over the 500-line
file budget and the largest LCM offender. Promote to a
``backend/app/core/lcm/evals/`` package per the project's
``package-multi-file-features`` rule. Five role-named modules + a
re-exporting ``__init__.py`` keep the public surface identical for every
existing call site.

Module layout:
- ``types.py`` — ``ScenarioIntent`` Literal, ``LCMEvalMode`` StrEnum,
  ``SeedSummary`` / ``LCMEvalScenario`` / ``LCMEvalResult`` dataclasses.
- ``seeding.py`` — ``seed_scenario``, ``_seed_turns``, ``_seed_summaries``,
  ``seed_embeddings_for_conversation``, ``_utcnow``, ``_approx_tokens``.
- ``answerer.py`` — ``deterministic_answer`` + ``_negative_answer``,
  ``_collect_pattern_quotes``, ``_fact_pass``, ``_source_pass``.
- ``retrievers.py`` — the seven ``_retrieve_*`` mode implementations
  (baseline / lcm_assembled / lcm_grep / lcm_search / hybrid /
  lcm_search_packed) plus ``_flatten_assembled`` and
  ``_extract_search_terms``.
- ``runner.py`` — the seven ``_adapt_*`` wrappers, ``_MODE_RETRIEVERS``
  dispatch table, ``_retrieve_for_mode``, ``run_eval``,
  ``run_eval_matrix``.

Public surface re-exported from ``evals/__init__.py``:
``ScenarioIntent``, ``LCMEvalMode``, ``SeedSummary``, ``LCMEvalScenario``,
``LCMEvalResult``, ``seed_scenario``,
``seed_embeddings_for_conversation``, ``deterministic_answer``,
``run_eval``, ``run_eval_matrix``. Verified by greppping every
``from app.core.lcm.evals import …`` site (just ``backend/tests/evals/``);
every imported name is in the list, no caller changes required.

Removed the ``backend/app/core/lcm/evals.py`` exemption from
``scripts/check-file-lines.mjs``. Updated the surrounding comment to
note ``embeddings.py`` is the only remaining LCM exemption (its split
lives on the sibling ``claude/lcm-embeddings-package`` branch / PR #386).

Updated ``backend/.importlinter``: the dead
``app.core.lcm.evals -> app.models`` entry is replaced by the two
concrete module-level edges that the package now creates —
``app.core.lcm.evals.seeding -> app.models`` and
``app.core.lcm.evals.retrievers -> app.models``.

Validation:
- ``node scripts/check-file-lines.mjs`` → OK (no files exceed 500 lines).
- ``cd backend && uv run mypy`` → 0 errors across 345 source files.
- ``cd backend && uv run ruff check .`` → clean.
- ``cd backend && uv run ruff format --check .`` → 346 files already formatted.
- ``cd backend && uv run lint-imports --config .importlinter`` →
  3 contracts kept, 0 broken.
- ``pytest tests/evals`` → 9 passed.
